requirements.txt
llama-cpp-python
huggingface-hub
mcp>=1.27.0
rich>=15.0.0
